move `PackageMatcher(s)` to package `core.domain`
This class does not really match the generic nature of classes in the `base` package, since it is clearly related to ArchUnit's domain. When some predicates were moved from `lang.conditions` to `domain.core` `PackageMatcher` was moved to `base`, probably out of the idea that then "every place within ArchUnit can utilize it". But in turn `PackageMatcher` can then not use any utilities from `core`, which makes no sense. Or all such utilities will start to creep into `base` as well. In the end, `core.domain` is the next higher layer on top of `base` in the ArchUnit layered architecture. And within `base` there should be no need ever to match packages, etc., since that is already ArchUnit domain specific logic. Since we are about to release 1.0 and can thus break the public API this is a good point in time to move this class back into a more reasonable layer.
