+++ archunit-example/example-plain/src/test/java/com/tngtech/archunit/exampletest/ControllerRulesTest.java
@@ -1,10 +1,10 @@
 package com.tngtech.archunit.exampletest;
 
 import com.tngtech.archunit.base.DescribedPredicate;
-import com.tngtech.archunit.base.PackageMatchers;
 import com.tngtech.archunit.core.domain.JavaClass;
 import com.tngtech.archunit.core.domain.JavaClasses;
 import com.tngtech.archunit.core.domain.JavaMember;
+import com.tngtech.archunit.core.domain.PackageMatchers;
 import com.tngtech.archunit.core.importer.ClassFileImporter;
 import com.tngtech.archunit.example.layers.security.Secured;
 import org.junit.Test;
