+++ archunit/src/main/java/com/tngtech/archunit/core/domain/PackageMatchers.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.tngtech.archunit.base;
+package com.tngtech.archunit.core.domain;
 
 import java.util.Collection;
 import java.util.Set;
@@ -21,6 +21,7 @@ import java.util.Set;
 import com.google.common.base.Joiner;
 import com.google.common.collect.ImmutableSet;
 import com.tngtech.archunit.PublicAPI;
+import com.tngtech.archunit.base.DescribedPredicate;
 
 import static com.tngtech.archunit.PublicAPI.Usage.ACCESS;
