+++ archunit/src/test/java/com/tngtech/archunit/core/domain/PackageMatchersTest.java
@@ -1,4 +1,4 @@
-package com.tngtech.archunit.base;
+package com.tngtech.archunit.core.domain;
 
 import org.junit.Test;
